refactor(command-bar): give the palette's state named hooks, and one page file
CommandPalette opened with twenty-one use_signal calls in a row. Nothing
said which belonged together, so reading any one of them meant scanning the
other twenty to find its pair — media_request_id and media_requested_query
guard each other, and neither reads as anything on its own.

Five hooks now, grouped by what the state is for: use_palette_input,
use_path_completions, use_history_suggestions, use_prompt_media and
use_prompt_target. Each returns a struct the component destructures under
the names the body already used, so the render is untouched. Where the
grouping made a field's job clear, it is written down — the request ids
exist because the host answers out of order and a late reply has to be
dropped rather than painted over newer rows.

palette.rs folds into page.rs. They were one page split across two files
for no reason but history: page.rs held the modal shell and rendered the
palette as its only child. Page leads the file now, CommandPalette follows
it, and the hooks and helpers follow that, so the file reads in the order
it composes.

The effects that drive this state have not moved into the hooks; they still
sit in the component. Doing that changes when they run, which is not
something a compile can check.
