+++ crates/page/vmux_layout/src/page.rs
@@ -17,7 +17,7 @@ use dioxus::html::input_data::MouseButton;
 use dioxus::prelude::*;
 use gloo_timers::future::TimeoutFuture;
 use vmux_command::event::CommandBarOpenEvent;
-use vmux_command::palette::{CommandPalette, PaletteVariant};
+use vmux_command::page::{CommandPalette, PaletteVariant};
 use vmux_core::event::team::{TEAM_EVENT, TeamCommandEvent, TeamEvent, TeamMemberRow};
 use vmux_core::event::{
     EXTENSIONS_LIST_EVENT, ExtActionRequest, ExtListRequest, ExtOpenManagerRequest, ExtRow,
